+++ src/Jet.ConfluentKafka.FSharp/ConfluentKafka.fs
@@ -52,10 +52,13 @@ type KafkaProducerConfig private (inner, broker : Uri) =
             /// Partition algorithm. Default: `ConsistentRandom`.
             ?partitioner,
             /// Miscellaneous configuration parameters to be passed to the underlying Confluent.Kafka producer configuration. Same as constructor argument for Confluent.Kafka >=1.2.
-            ?custom : IDictionary<string,string>,
+            ?config : IDictionary<string,string>,
+            /// Miscellaneous configuration parameters to be passed to the underlying Confluent.Kafka producer configuration.
+            ?custom,
+            /// Postprocesses the ProducerConfig after the rest of the rules have been applied
             ?customize) =
         let c =
-            let customPropsDictionary = match custom with Some x -> x | None -> Dictionary<string,string>() :> IDictionary<string,string>
+            let customPropsDictionary = match config with Some x -> x | None -> Dictionary<string,string>() :> IDictionary<string,string>
             ProducerConfig(customPropsDictionary, // CK 1.2 and later has a default ctor and an IDictionary<string,string> overload
                 ClientId = clientId, BootstrapServers = Config.validateBrokerUri broker,
                 RetryBackoffMs = Nullable (match retryBackoff with Some (t : TimeSpan) -> int t.TotalMilliseconds | None -> 1000), // CK default 100ms
@@ -68,6 +71,7 @@ type KafkaProducerConfig private (inner, broker : Uri) =
         partitioner |> Option.iter (fun x -> c.Partitioner <- Nullable x)
         compression |> Option.iter (fun x -> c.CompressionType <- Nullable x)
         statisticsInterval |> Option.iter<TimeSpan> (fun x -> c.StatisticsIntervalMs <- Nullable (int x.TotalMilliseconds))
+        custom |> Option.iter (fun xs -> for KeyValue (k,v) in xs do c.Set(k,v))
         customize |> Option.iter (fun f -> f c)
         KafkaProducerConfig(c, broker)
 
@@ -205,8 +209,10 @@ type KafkaConsumerConfig = private { inner: ConsumerConfig; topics: string list;
             ?statisticsInterval,
             /// Consumed offsets commit interval. Default 5s.
             ?offsetCommitInterval,
-            /// Misc configuration parameters to be passed to the underlying CK consumer. (Same as constructor argument for Confluent.Kafka >=1.2.
-            ?custom : IDictionary<string,string>,
+            /// Misc configuration parameters to be passed to the underlying CK consumer. Same as constructor argument for Confluent.Kafka >=1.2.
+            ?config : IDictionary<string,string>,
+            /// Misc configuration parameter to be passed to the underlying CK consumer.
+            ?custom,
             /// Postprocesses the ConsumerConfig after the rest of the rules have been applied
             ?customize,
 
@@ -224,7 +230,7 @@ type KafkaConsumerConfig = private { inner: ConsumerConfig; topics: string list;
         let minInFlightBytes = defaultArg minInFlightBytes (maxInFlightBytes * 2L / 3L)
         let fetchMaxBytes = defaultArg fetchMaxBytes 100_000
         let c =
-            let customPropsDictionary = match custom with Some x -> x | None -> Dictionary<string,string>() :> IDictionary<string,string>
+            let customPropsDictionary = match config with Some x -> x | None -> Dictionary<string,string>() :> IDictionary<string,string>
             ConsumerConfig(customPropsDictionary, // CK 1.2 and later has a default ctor and an IDictionary<string,string> overload
                 ClientId=clientId, BootstrapServers=Config.validateBrokerUri broker, GroupId=groupId,
                 AutoOffsetReset = Nullable (defaultArg autoOffsetReset AutoOffsetReset.Earliest), // default: latest
@@ -236,6 +242,7 @@ type KafkaConsumerConfig = private { inner: ConsumerConfig; topics: string list;
         fetchMinBytes |> Option.iter (fun x -> c.FetchMinBytes <- x) // Fetch waits for this amount of data for up to FetchWaitMaxMs (100)
         offsetCommitInterval |> Option.iter<TimeSpan> (fun x -> c.AutoCommitIntervalMs <- Nullable <| int x.TotalMilliseconds)
         statisticsInterval |> Option.iter<TimeSpan> (fun x -> c.StatisticsIntervalMs <- Nullable <| int x.TotalMilliseconds)
+        custom |> Option.iter (fun xs -> for KeyValue (k,v) in xs do c.Set(k,v))
         customize |> Option.iter<ConsumerConfig -> unit> (fun f -> f c)
         {   inner = c
             topics = match Seq.toList topics with [] -> invalidArg "topics" "must be non-empty collection" | ts -> ts